Implement `keep_complete_c_asm_splits_for_objdiff_compatibility`
